endpoint_adapters/requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
pika==1.3.1
platformdirs==2.6.0
praw==7.6.1
prawcore==2.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
